fix(artifacts/k8s): fix rewrite k8s artifact edit (#7352)
the match artifact editor used for kubernetes match artifacts wasn't
working right. previously, you were only presented with a single
reference field which didn't capture all of the data we needed to match
on artifacts. instead of something like `kubernetes/configMap` the
default artifact type was `kubernetes/REFERENCE` which is not a valid
type but there was no way to change it.