fix(core/infrastructure): Fix deep links with filters (#7459)
This fixes a problem where deep links to infrastructure items (such as an instance chiclet) where filters are applied will lose the filters when the deep link is activated.  This was due to a but introduced when refactoring the filter save/restore code a few months back.  We started using router hooks to manage saving/restoring the filters.  However, the case where a user was navigating deeply was not accounted for in the router trigger criteria.  It only took effect when the top level view was activated.

This change switches to triggering the router hook on a wildcard glob on the beginning and end of the state names.
This causes the AngularJS compatibility code to be run whether you're activating the clusters tab, or deep linking to an instance _within_ the clusters tab.