fix(core/confirmationModal): Supply the 'source' location when canceling (#7816)
This fixes a bug in Enable Server Group modal.
When the user selects the "No, I just want to enable the server group", it was closing the dialog and doing nothing else.
Now, it closes the first dialog and opens the next dialog (but not if they click the X in the top right corner)

Also, fixed the boolean logic that shows the "Reason" input.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>