feat(codebuild): Add support for multiple sources (#7915)
* feat(codebuild): Add support for multiple sources

Refactor the stage definition so that all source related fields go into one single structure.
This is helpful when it comes to multiple sources because of the reusability that the single
structure provides. For primary source, artifact is not required. Users can override source
version without touching source artifact. However, in secondary sources, users must specify
an artifact. Otherwise, there's no way to link a secondary source version to a secondary source
defined in the project configuration.

* refactor(codebuild): Create a generic function to update field

Co-authored-by: Clare Liguori <liguori@amazon.com>