feat(core): enable standard artifacts UI by default (#8184)
Related to: https://github.com/spinnaker/governance/blob/master/rfc/legacy_artifacts_ui_removal.md,

This change removes `artifacts` and `artifactsRewrite` feature flags in preparation for release 1.20, in which the standard artifacts UI will be enabled by default. Provides a temporary flag that can be enabled in order to revert to the legacy UI, which will be removed in release 1.21.